+++ pmd/src/site/markdown/changelog.md
@@ -4,3 +4,6 @@
 
 **Bugfixes:**
 
+* Fixed [bug #1192]: Ecmascript fails to parse this operator " ^= "
+
+[bug #1192]: https://sourceforge.net/p/pmd/bugs/1192/